+++ packages/react/src/components/UIShell/components/TrialCountdown.tsx
@@ -1,5 +1,5 @@
 /**
- * Copyright IBM Corp. 2025
+ * Copyright IBM Corp. 2025, 2026
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -40,7 +40,7 @@ export const TrialCountdown: React.FC<TrialCountdownProps> = ({
   });
   return (
     <p className={classNames}>
-      <span className={`${prefix}--trial-countdown__count`}>{count}</span>{' '}
+      <span className={`${prefix}--trial-countdown__count`}>{count}</span>
       {text}
     </p>
   );