+++ packages/web-components/src/components/global-header/components/global-header/src/components/HybridIpaasHeader/__codecept__/HybridIpaasHeader.spec.js
@@ -172,8 +172,11 @@ Scenario('Solis components render', async ({ I }) => {
 
   I.seeElement(locate('#ibm-automation-cds-solis-switcher-button'));
   I.click(locate('#ibm-automation-cds-solis-switcher-button'));
-  I.see('Observability');
+  // I.see('Observability');
+  I.wait(2);
+  I.see('Your products');
   I.see('Community');
   I.click(locate('#ibm-automation-cds-solis-switcher-button')); // Close the Solis switcher
-  I.dontSee('Observability'); // Solis switcher is closed
+  // I.dontSee('Observability'); // Solis switcher is closed
+  I.dontSee('Your products');
 });